Move withRouteApplicationState HOC into spaces/application area
