+++ Source/SelfService/Web/spaces/applications/withRouteApplicationState.tsx
@@ -2,8 +2,8 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 import React, { useEffect, ComponentType } from 'react';
-import { useGlobalContext } from '../context/globalContext';
-import { RouteApplicationParams, useRouteApplicationParams } from '../utils/route';
+import { useGlobalContext } from '../../context/globalContext';
+import { RouteApplicationParams, useRouteApplicationParams } from '../../utils/route';
 
 export type WithRouteApplicationProps = {
     routeApplicationParams: RouteApplicationParams
